+++ wezterm-input-types/src/lib.rs
@@ -1038,6 +1038,16 @@ fn normalize_shift(key: KeyCode, modifiers: Modifiers) -> (KeyCode, Modifiers) {
     }
 }
 
+pub fn is_ascii_control(c: char) -> Option<char> {
+    let c = c as u32;
+    if c < 0x20 {
+        let de_ctrl = ((c as u8) | 0x40) as char;
+        Some(de_ctrl.to_ascii_lowercase())
+    } else {
+        None
+    }
+}
+
 fn normalize_ctrl(key: KeyCode, modifiers: Modifiers) -> (KeyCode, Modifiers) {
     if modifiers.contains(Modifiers::CTRL) {
         if let KeyCode::Char(c) = key {